+++ js/src/index.ts
@@ -1,4 +1,4 @@
 export { interiorPoint } from "./algorithm/InteriorPoint.ts";
 export { centroidFirstInteriorPoint } from "./CentroidFirstInteriorPoint.ts";
-export { verifyInteriorPoint, InteriorPointVerification } from "./VerifyInteriorPoint.ts";
+export { verifyInteriorPoint, Verification } from "./VerifyInteriorPoint.ts";
 export type { Coordinate } from "./GeometryAdapter.ts";